+++ commands/help.js
@@ -1,0 +1,13 @@
+const fs = require('fs')
+const path = require('path')
+
+function command (args, flags, context) {
+  
+}
+
+const options = []
+
+module.exports = {
+  command,
+  options
+}